backend/requirements.txt
Django>=4.2
mysqlclient>=2.1
python-dotenv>=1.0
django-cors-headers>=4.0
